fix: audit complet owner — safelist Tailwind v4, analytics DB query, approved status, vnc-bootstrap + post-deploy
- resources/css/app.css: @source inline() safelist pour toutes les couleurs dynamiques (emerald/amber/pink…)
- AnalyticsController: firstResidenceId extrait côté controller (plus de DB query dans la vue)
- analytics/index.blade.php: remplacé auth()->user()->residences()->first() par $firstResidenceId
- residences/index.blade.php: 3 fixes status 'approved' manquant
- residences/show.blade.php: fix status 'approved' manquant
- residences/wizard.blade.php + ResidenceController: Amenity::all() déplacé en controller
- dashboard.blade.php: 6 fixes (statusConfig approved, onboarding, totalOwners, latestMessage)
- statistics.blade.php: 2 fixes (approved badge, dynamic Tailwind class)
- OwnerController: 4 fixes (User import, eager load latestMessage, totalOwners cache)
- vnc-bootstrap.sh: fix ordre UFW, supervisor démarrage, cron www-data, storage/logs mkdir
- post-deploy.sh: créé (composer prod, migrate, cache, permissions, queue restart)

+++ app/Models/PlatformSetting.php
@@ -88,7 +88,7 @@ class PlatformSetting extends Model
      */
     public static function getCommissionRate(): float
     {
-        return static::getValue('commission_rate', 10);
+        return (float) static::getValue('commission_rate', 10);
     }
 
     /**
@@ -97,7 +97,7 @@ class PlatformSetting extends Model
     public static function calculateCommission(float $amount): float
     {
         $rate = static::getCommissionRate();
-        $min = static::getValue('commission_min', 1000);
+        $min = (float) static::getValue('commission_min', 1000);
 
         $commission = $amount * ($rate / 100);
